abstract providers even more to get rid of MCWorld usage in Chunk.java in the server
