--- pizzaserver-mcworld/pom.xml
@@ -1,49 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <artifactId>PizzaServer</artifactId>
-        <groupId>io.github.willqi.pizzaserver</groupId>
-        <version>1.0-SNAPSHOT</version>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-
-    <groupId>io.github.willqi.pizzaserver.mcworld</groupId>
-    <artifactId>PizzaServer-MCWorld</artifactId>
-
-    <properties>
-        <maven.compiler.source>8</maven.compiler.source>
-        <maven.compiler.target>8</maven.compiler.target>
-    </properties>
-
-    <repositories>
-        <repository>
-            <id>opencollab-repo-snapshot</id>
-            <url>https://repo.opencollab.dev/maven-snapshots/</url>
-        </repository>
-    </repositories>
-
-    <dependencies>
-
-        <dependency>
-            <groupId>io.github.willqi.pizzaserver.server.nbt</groupId>
-            <artifactId>PizzaServer-NBT</artifactId>
-            <version>1.0-SNAPSHOT</version>
-        </dependency>
-        <dependency>
-            <groupId>io.github.willqi.pizzaserver.commons</groupId>
-            <artifactId>PizzaServer-Commons</artifactId>
-            <version>1.0-SNAPSHOT</version>
-        </dependency>
-
-        <!-- LevelDB in Java that supports MCPE format -->
-        <dependency>
-            <groupId>net.daporkchop</groupId>
-            <artifactId>leveldb-mcpe-jni</artifactId>
-            <version>0.0.7-SNAPSHOT</version>
-        </dependency>
-
-    </dependencies>
-
-</project>
