--- pizzaserver-mcworld/src/main/java/io/github/willqi/pizzaserver/mcworld/BlockRuntimeMapper.java
@@ -1,9 +1,0 @@
-package io.github.willqi.pizzaserver.mcworld;
-
-import io.github.willqi.pizzaserver.nbt.tags.NBTCompound;
-
-public interface BlockRuntimeMapper {
-
-    int getBlockRuntimeId(String name, NBTCompound states);
-
-}
