--- pizzaserver-mcworld/src/main/java/io/github/willqi/pizzaserver/mcworld/world/chunks/subchunks/BedrockSubChunk.java
@@ -1,103 +1,0 @@
-package io.github.willqi.pizzaserver.mcworld.world.chunks.subchunks;
-
-import io.github.willqi.pizzaserver.mcworld.BlockRuntimeMapper;
-import io.netty.buffer.ByteBuf;
-import io.netty.buffer.ByteBufAllocator;
-
-import java.io.IOException;
-import java.io.UnsupportedEncodingException;
-import java.util.ArrayList;
-import java.util.List;
-
-
-public class BedrockSubChunk {
-
-    private int subChunkVersion;
-    private final List<BlockLayer> layers = new ArrayList<>();
-
-
-    public int getSubChunkVersion() {
-        return this.subChunkVersion;
-    }
-
-    public BlockLayer getLayer(int layer) {
-        return this.layers.get(layer);
-    }
-
-    public List<BlockLayer> getLayers() {
-        return this.layers;
-    }
-
-    public void addLayer(BlockLayer layer) {
-        this.layers.add(layer);
-    }
-
-    /**
-     * Construct a sub chunk based on data retrieved from the level database.
-     * @param buffer
-     * @throws IOException
-     */
-    public void parse(ByteBuf buffer) throws IOException {
-        this.subChunkVersion = buffer.readByte();
-        int totalLayers;
-
-        switch (subChunkVersion) {
-            case 8:
-                totalLayers = buffer.readByte();
-                break;
-            case 1:
-                totalLayers = 1;
-                break;
-            default:
-                throw new UnsupportedEncodingException("Missing implementation for v" + subChunkVersion + " chunks");
-        }
-
-        for (int layerI = 0; layerI < totalLayers; layerI++) {
-            int bitsPerBlock = buffer.readByte() >> 1;
-            int blocksPerWord = 32 / bitsPerBlock;
-            int wordsPerChunk = (int)Math.ceil(4096d / blocksPerWord);
-
-            // We want to read the palette first so we can translate what blocks are immediately.
-            int chunkBlocksIndex = buffer.readerIndex();
-            buffer.setIndex(chunkBlocksIndex + (wordsPerChunk * 4), buffer.writerIndex());
-
-            BlockPalette palette = new BlockPalette();
-            palette.parse(buffer);
-            int endPaletteIndex = buffer.readerIndex(); // we jump to this index after reading the blocks
-
-            // Go back and parse the blocks.
-            buffer.setIndex(chunkBlocksIndex, buffer.writerIndex());
-            BlockLayer layer = new BlockLayer(palette);
-            layer.parse(buffer, bitsPerBlock, blocksPerWord, wordsPerChunk);
-            this.layers.add(layer);
-
-            // Go back to the end of the palette to prepare for the next layer
-            buffer.setIndex(endPaletteIndex, buffer.writerIndex());
-        }
-    }
-
-    public byte[] serializeForDisk() throws IOException {
-        ByteBuf buffer = ByteBufAllocator.DEFAULT.buffer(1);
-        buffer.writeByte(8);    // Convert to version 8 regardless of v1 or v8
-        buffer.writeByte(this.layers.size());
-        this.layers.forEach(layer -> buffer.writeBytes(layer.serializeForDisk()));
-
-        byte[] serialized = new byte[buffer.readableBytes()];
-        buffer.readBytes(serialized);
-        buffer.release();
-        return serialized;
-    }
-
-    public byte[] serializeForNetwork(BlockRuntimeMapper runtimeMapper) throws IOException {
-        ByteBuf buffer = ByteBufAllocator.DEFAULT.buffer(1);
-        buffer.writeByte(8);    // Convert to version 8 regardless of v1 or v8
-        buffer.writeByte(this.layers.size());
-        this.layers.forEach(layer -> buffer.writeBytes(layer.serializeForNetwork(runtimeMapper)));
-
-        byte[] serialized = new byte[buffer.readableBytes()];
-        buffer.readBytes(serialized);
-        buffer.release();
-        return serialized;
-    }
-
-}
