+++ pizzaserver-server/pom.xml
@@ -37,8 +37,8 @@
             <version>1.0-SNAPSHOT</version>
         </dependency>
         <dependency>
-            <groupId>io.github.willqi.pizzaserver.mcworld</groupId>
-            <artifactId>PizzaServer-MCWorld</artifactId>
+            <groupId>io.github.willqi.pizzaserver.format</groupId>
+            <artifactId>PizzaServer-Format</artifactId>
             <version>1.0-SNAPSHOT</version>
         </dependency>
