+++ pizzaserver-worldfile/src/main/java/io/github/willqi/pizzaserver/format/mcworld/BlockRuntimeMapper.java
@@ -1,0 +1,9 @@
+package io.github.willqi.pizzaserver.format.mcworld;
+
+import io.github.willqi.pizzaserver.nbt.tags.NBTCompound;
+
+public interface BlockRuntimeMapper {
+
+    int getBlockRuntimeId(String name, NBTCompound states);
+
+}
